push current state of software

+++ src/main/java/com/kristian/dimitrov/HyperlapseRobot/hardware/ContinuousRotationServoComponent.java
@@ -114,7 +114,7 @@ public class ContinuousRotationServoComponent {
         return pwm.getFrequency();
     }
 
-    public void setClockwiseRotationDutyCycle(int dutyCycle){
+    public void setClockwiseRotationDutyCycle(float dutyCycle){
         this.clockwise_rot_duty_cycle = dutyCycle;
     }
 }
