+++ src/main/java/com/kristian/dimitrov/HyperlapseRobot/service/ContinuousRotServoService.java
@@ -2,6 +2,6 @@ package com.kristian.dimitrov.HyperlapseRobot.service;
 
 public interface ContinuousRotServoService {
     void stop();
-    void rotateClockwise(int frequency, int clockwiseRotationDutyCycle);
+    void rotateClockwise(int frequency, float clockwiseRotationDutyCycle);
     void rotateCounterClockwise(int frequency);
 }
